Utilização do inner class na validação dos produtos.
